Move getHelplineToSave out of formSubmissionHelpers
